Assignment 07 Custom Web Search Tool/pyproject.toml
[project]
name = "assignment-07-custom-web-search-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
    "tavily-python>=0.7.11",
]
